Still working on tests. Refactored out outbound messages to separate file
